fixed datatypes and test

+++ src/DataTypes.java
@@ -4,9 +4,9 @@ public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
-        for (int x : numbers) {
+        for (Integer x : numbers) {
             s += x;
         }
         return s;
